entity fields configuration is now a part of EntityConfig
